Make the feedback logic DI
